+++ magda-web-client/src/Components/Dataset/View/DatasetPageDetails.js
@@ -8,6 +8,7 @@ import defined from "helpers/defined";
 import CommonLink from "Components/Common/CommonLink";
 import DiscourseComments from "Components/Dataset/View/DiscourseComments";
 import { config } from "../../../config";
+import MagdaNamespacesConsumer from "Components/i18n/MagdaNamespacesConsumer";
 import "./DatasetDetails.scss";
 import "./DatasetPageDetails.scss";
 
@@ -22,93 +23,109 @@ class DatasetPageDetails extends Component {
         const sourceName = this.props?.dataset?.sourceDetails?.originalName
             ? this.props.dataset.sourceDetails.originalName
             : this.props.dataset?.sourceDetails?.name;
-        const source = sourceName
-            ? `This dataset was originally found on ${
-                  dataset.landingPage
-                      ? `[${sourceName}](${dataset.landingPage})`
-                      : sourceName
-              } "${dataset.title}".${
-                  dataset.landingPage
-                      ? "\nPlease visit the source to access the original metadata of the dataset:"
-                      : ""
-              }`
-            : "";
+        const sourceId = this.props?.dataset?.sourceDetails?.id;
         return (
-            <div className="dataset-details">
-                <div className="dataset-preview">
-                    <DatasetPagePreview
-                        location={this.props.location}
-                        dataset={dataset}
-                    />
-                </div>
-
-                <div className="row">
-                    <div className="col-sm-12">
-                        <div className="dataset-details-files-apis">
-                            <h3 className="clearfix section-heading">
-                                <span className="section-heading">
-                                    Files and APIs
-                                </span>
-                            </h3>
-                            <div className="clearfix">
-                                {dataset.distributions.map((s) => (
-                                    <DistributionRow
-                                        key={s.identifier}
-                                        distribution={s}
-                                        dataset={dataset}
-                                        searchText={searchText}
-                                    />
-                                ))}
+            <MagdaNamespacesConsumer ns={["global"]}>
+                {(translate) => {
+                    const appName = translate(["appName", ""]);
+                    const source =
+                        sourceId === "magda"
+                            ? `This dataset was created internally by ${appName}.`
+                            : sourceName
+                            ? `This dataset was originally found on ${
+                                  dataset.landingPage
+                                      ? `[${sourceName}](${dataset.landingPage})`
+                                      : sourceName
+                              } "${dataset.title}".${
+                                  dataset.landingPage
+                                      ? "\nPlease visit the source to access the original metadata of the dataset:"
+                                      : ""
+                              }`
+                            : "";
+                    return (
+                        <div className="dataset-details">
+                            <div className="dataset-preview">
+                                <DatasetPagePreview
+                                    location={this.props.location}
+                                    dataset={dataset}
+                                />
                             </div>
-                        </div>
-                        {(source || dataset.provenance) && (
-                            <div className="dataset-details-source">
-                                <h3 className="section-heading">Data Source</h3>
-                                {source && (
-                                    <MarkdownViewer
-                                        markdown={source}
-                                        truncate={false}
-                                    />
-                                )}
-                                {dataset?.landingPage ? (
-                                    <CommonLink
-                                        className="landing-page"
-                                        href={dataset.landingPage}
-                                    >
-                                        {dataset.landingPage}
-                                    </CommonLink>
-                                ) : null}
 
-                                {defined(dataset.provenance) &&
-                                defined(dataset.provenance.isOpenData) ? (
-                                    <h3 className="section-heading">
-                                        Type:{" "}
-                                        {dataset.provenance.isOpenData
-                                            ? "Public"
-                                            : "Private"}
-                                    </h3>
-                                ) : null}
+                            <div className="row">
+                                <div className="col-sm-12">
+                                    <div className="dataset-details-files-apis">
+                                        <h3 className="clearfix section-heading">
+                                            <span className="section-heading">
+                                                Files and APIs
+                                            </span>
+                                        </h3>
+                                        <div className="clearfix">
+                                            {dataset.distributions.map((s) => (
+                                                <DistributionRow
+                                                    key={s.identifier}
+                                                    distribution={s}
+                                                    dataset={dataset}
+                                                    searchText={searchText}
+                                                />
+                                            ))}
+                                        </div>
+                                    </div>
+                                    {(source || dataset.provenance) && (
+                                        <div className="dataset-details-source">
+                                            <h3 className="section-heading">
+                                                Data Source
+                                            </h3>
+                                            {source && (
+                                                <MarkdownViewer
+                                                    markdown={source}
+                                                    truncate={false}
+                                                />
+                                            )}
+                                            {dataset?.landingPage ? (
+                                                <CommonLink
+                                                    className="landing-page"
+                                                    href={dataset.landingPage}
+                                                >
+                                                    {dataset.landingPage}
+                                                </CommonLink>
+                                            ) : null}
+
+                                            {defined(dataset.provenance) &&
+                                            defined(
+                                                dataset.provenance.isOpenData
+                                            ) ? (
+                                                <h3 className="section-heading">
+                                                    Type:{" "}
+                                                    {dataset.provenance
+                                                        .isOpenData
+                                                        ? "Public"
+                                                        : "Private"}
+                                                </h3>
+                                            ) : null}
+                                        </div>
+                                    )}
+                                </div>
                             </div>
-                        )}
-                    </div>
-                </div>
 
-                {config.discourseSiteUrl &&
-                config.discourseIntegrationDatasetPage ? (
-                    <div className="row">
-                        <div className="col-sm-12">
-                            <DiscourseComments
-                                title={
-                                    dataset?.title
-                                        ? dataset.title
-                                        : "Untitlted Dataset"
-                                }
-                                datasetId={dataset.identifier}
-                            />
+                            {config.discourseSiteUrl &&
+                            config.discourseIntegrationDatasetPage ? (
+                                <div className="row">
+                                    <div className="col-sm-12">
+                                        <DiscourseComments
+                                            title={
+                                                dataset?.title
+                                                    ? dataset.title
+                                                    : "Untitlted Dataset"
+                                            }
+                                            datasetId={dataset.identifier}
+                                        />
+                                    </div>
+                                </div>
+                            ) : null}
                         </div>
-                    </div>
-                ) : null}
-            </div>
+                    );
+                }}
+            </MagdaNamespacesConsumer>
         );
     }
 }